refactor: 영화 예매 결과 화면 개선
