+++ app/src/main/java/woowacourse/movie/presentation/view/reservation/result/ReservationResultPresenter.kt
@@ -1,20 +1,14 @@
 package woowacourse.movie.presentation.view.reservation.result
 
-import woowacourse.movie.domain.model.ReservationInfo
-import woowacourse.movie.presentation.model.ReservationInfoUiModel
-import woowacourse.movie.presentation.model.toModel
+import woowacourse.movie.domain.model.cinema.TicketMachine
+import woowacourse.movie.presentation.model.TicketBundleUiModel
 
 class ReservationResultPresenter(
     private val view: ReservationResultContract.View,
 ) : ReservationResultContract.Presenter {
-    private var reservationInfo: ReservationInfo? = null
-
-    val cancellationTime = ReservationInfo.CANCELLATION_TIME
-
-    override fun fetchDate(getReservationInfo: () -> ReservationInfoUiModel?) {
-        getReservationInfo()?.let { info ->
-            reservationInfo = info.toModel()
-            view.setScreen(info)
+    override fun fetchDate(ticketBundle: TicketBundleUiModel?) {
+        ticketBundle?.let { bundle ->
+            view.setScreen(bundle, TicketMachine.CANCELLATION_TIME)
             return
         }
