+++ app/src/test/java/woowacourse/movie/presentation/reservation/result/ReservationResultPresenterTest.kt
@@ -7,8 +7,8 @@ import io.mockk.mockk
 import io.mockk.verify
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
-import woowacourse.movie.domain.model.ReservationCount
-import woowacourse.movie.domain.model.ReservationInfo
+import woowacourse.movie.domain.model.reservation.ReservationCount
+import woowacourse.movie.domain.model.reservation.ReservationInfo
 import woowacourse.movie.presentation.model.toUiModel
 import woowacourse.movie.presentation.view.reservation.result.ReservationResultContract
 import woowacourse.movie.presentation.view.reservation.result.ReservationResultPresenter
